+++ CHANGELOG.md
@@ -5,6 +5,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 Nothing
 
+## [1.0.0] - 2022-05-01
+### Added
+- Add setting to auto-start Taskbar Notifier
+- Allow changing the notification location
+- Allow closing the notification with any mouse event
+- Add setting to flash the screen shortly on notifications
+
 ## [0.3.0] - 2020-12-08
 ### Changed
 - Replace Windows toast notifications with custom ones
@@ -23,7 +30,8 @@ Nothing
 ## [0.1.0] - 2018-12-08
 First release
 
-[Unreleased]: https://github.com/rfkd/TaskbarNotifier/compare/0.3.0...HEAD
+[Unreleased]: https://github.com/rfkd/TaskbarNotifier/compare/1.0.0...HEAD
+[1.0.0]: https://github.com/rfkd/TaskbarNotifier/compare/0.3.0...1.0.0
 [0.3.0]: https://github.com/rfkd/TaskbarNotifier/compare/0.2.1...0.3.0
 [0.2.1]: https://github.com/rfkd/TaskbarNotifier/compare/0.2.0...0.2.1
 [0.2.0]: https://github.com/rfkd/TaskbarNotifier/compare/0.1.0...0.2.0